nvidia: add DMI blacklist
We have found multiple systems where the nvidia driver introduces
significant regressions over the free drivers.

We intend to work with nvidia to fix these issues, but need a solution
while that is in process.

Add a DMI blacklist mechanism, and populate it with our first 2 platforms
where this must apply.

The user can override this blacklist by providing their own
in /etc/endless-external-drivers/nvidia/dmi-blacklist and
creating this as an empty file will basically cause the ostree-shipped
blacklist to be ignored.

https://phabricator.endlessm.com/T18554