add eos-extra-target-generator to create eos-extra-settled.target
Introduce a new systemd.generator called eos-extra-target-generator, to address
race conditions when running eos-update-flatpak-repos on "split" systems where
the default flatpak system dir is provided by an SD card mounted at
/var/endless-extra. The problem is that (by design) the local-fs.target is not
blocked by the availability of /var/endless-extra so that systems can boot
properly when the SD card is unavailable - whether missing, damaged or
corrupted - the only symptom is the availability of the apps installed there,
but the rest of the system works OK.

If the discovery of the SD card is slow enough, or it's being resized or
fscked, the /var/endless-extra mount job isn't in the queue when the
eos-update-flatpak-repos job is evaluated, so the After= has no effect and the
job runs before /var/endless-extra is mounted, causing flatpak to operate on
/var/lib/flatpak (it selects /var/endless-extra by runtime detection of
/var/endless-extra/flatpak).

The fix is to introduce a new target whose conditions are met either only after
/var/endless-extra is mounted, or immediately if the system does not expect to
have an SD card. The target is created at boot time by
eos-extra-target-generator, which checks for the link at
/etc/systemd/system/local-fs.wants created by the image builder to enable
/var/endless-extra on split systems, and adds RequiresMountFor if appropriate.

The eos-update-flatpak-repos and eos-fix-flatpak-branches units are updated to
require eos-extra-settled.target. Note that this target should never be used to
do critical work on paths outside of the SD card, because in the case the SD
card is unavailable on a split system, those would never run.

https://phabricator.endlessm.com/T20443