Add eos-grub-upgrade and related service file
Add the new script and related service file implementing the mechanism
to upgrade GRUB and SHIM files in the ESP.

Right now the updater is only for single-boot UEFI platforms.

The script checks, only after a new ostree is deployed, if the files
present in the ESP (SHIM and GRUB) need to be upgraded by comparing
these files with the corresponding files in fixed location in the
rootfs. The comparison is done by checking if the files are different
and if the files are signed or not.

If the script determines that the files need to be updated, the new
files are copied in the same partition as the old ones and a 'mv` is
issued to try to be as atomic as possible.

Signed-off-by: Carlo Caione <carlo@endlessm.com>