Add eos-boot-upgrade and related service file
Add the new script and related service file implementing the mechanism
to upgrade GRUB, SHIM and boot config files.

Right now the updater is only for single-boot UEFI platforms.

The script checks, only after a new ostree is deployed, if the files
present in the ESP (SHIM and GRUB) and the configuration files need to
be upgraded. If that's the case the new files are copied from the ostree
deploy in the correct place trying to be as much atomic as possible when
replacing the old files.

Signed-off-by: Carlo Caione <carlo@endlessm.com>