50-eos.preset: Disable apt-daily{,-upgrade}.timer
On ostree systems, the associated .service units will always fail when
triggered. So, arrange for them to be disabled by default.

It is a little sad that this applies to converted systems where these
timers actually are useful. An alternative would be to add

    ConditionKernelCommandLine=!ostree

to apt-daily{,upgrade}.{timer,service} in the 'apt' package (or as
drop-ins in UNIT.d/ostree.conf files) but as discussed in
https://phabricator.endlessm.com/T18831#470667 this would impose more
maintenance overhead in future unless we were to push this change
upstream.

https://phabricator.endlessm.com/T18831