endless-repartition.sh: Allow multiple part attrs on rootfs
At least for the rk3288 platform the rootfs partition is marked with
multiple attrs, in particular:

LegacyBIOSBootable GUID:55

In the current endless repartiton code we are only matching the GUID:55
string, making it failing when multiple attributes are present.

Relax a bit the string comparison, allowing GUID:55 to be just one of
the rootfs partition attributes.

Signed-off-by: Carlo Caione <carlo@endlessm.com>