repartition: Don't use backslash in sed pattern
sed is not matching the sysroot device unit name when we try to update
it because it contains a backslash (\). Since we know there will be only
one device unit and it will always be identified by UUID, we can use a
regular expression to match it.

Also, lets call systemd-escape with the -p switch to make it ignore the
leading slash from the input string (not sure why this worked before)
and pass --suffix=device so it generates a full device unit name.

https://phabricator.endlessm.com/T21532