dmi-blacklist: ASUS FX570UD & V272UN can't using nv drv
The NV GTX 1050M of FX570UD and NV MX150 of V272UN cannot resume back
from suspend with the normal display when use the nvidia driver (ver. 390.25).
If we ssh into the system, we can find that X takes the CPU resource almost
100%.
So, change the driver to nouveau which can solve the problem.

https://phabricator.endlessm.com/T21413