endless-repartition.sh: Preserve part attrs on UUID randomization
When repartitioning on first boot we are randomizing the UUID by
removing the UUID string from the 'sfdisk' output. When doing so we are
also wrongly removing the partition attributes reported by 'sfdisk'
immediately after the UUID. We need to keep these attributes because on
the chromebook (rk3288 image) they are interpreted as Chromium OS
extensions to the GPT partitions and used by depthcharge especially
during the boot phase.

We slightly rework the code used to remove the UUID from 'sfdisk' output
taking care of only removing the UUID string preserving the partition
attributes.

Signed-off-by: Carlo Caione <carlo@endlessm.com>