endless-repartition.sh: Carefully remove only the marker
At the end of repartition we want to remove the GUID:55 marker from the
GPT partitions.

Unfortunately in the current code we are removing all the partition
attributes, not only the marker. In the rk3288 platform the
'LegacyBIOSBootable' attribute is needed to identify the booting
partition and we need to preserve it.

Fix this by only removing the GUID:55 marker without touching the
remaining attributes.

Signed-off-by: Carlo Caione <carlo@endlessm.com>