endless-repartition.sh: Carefully remove only the marker
At the end of repartition we want to remove the GUID:55 marker from the
GPT partitions.

Unfortunately in the current code we are removing all the partition
attributes, not only the marker. In the rk3288 platform the
'LegacyBIOSBootable' attribute is needed to identify the booting
partition and we need to preserve it.

Fix this by only removing the GUID:55 marker without touching the
remaining attributes.

For the rk3288 platform the rootfs partition is marked with (output of
'sfdisk --part-attrs ...':

LegacyBIOSBootable GUID:55

For other platforms we only have:

GUID:55

We need to delete 'GUID:55' without touching the 'LegacyBIOSBootable'
attribute used by U-Boot to identify the partition containing the /boot
directory.

Signed-off-by: Carlo Caione <carlo@endlessm.com>