Disable disk swap and enable zram for all users
Performance is significantly degraded when swapping to spinning HDDs,
and zram offers much better performance.  For now, let's disable
the swap partition at boot for all users and use zram instead.
In the future, we may consider other options such as zswap to make
more effective use of disk swapping, or combining zram with swap
(if we can ensure that zram has the higher priority), or removing
the swap partition altogether.

https://phabricator.endlessm.com/T21726