eos-update-flatpak-repos: remove broken Flatpak remotes.d configuration
When we started dynamically creating Flatpak remotes on upgrades using
flatpak remote-add in boot scripts, we uncovered a bug in ostree's remote
configuration that meant ostree was creating entries in /etc/ostree/remotes.d
but flatpak was directly editing /ostree/repo/config, meaning that the
remote was created twice. This was fixed around 3.3.0 by setting
add-remotes-config-dir to false in an init script. (this was T19077)

People who upgraded before this fix could still have broken entries in
remotes.d, so we forcibly remove these files and any config in the ostree
repo, and then allow the remotes to be dynamically re-added. The OS
(including this script, and more critically now the OS updater) relies on
these remotes being present for upgrades to take place.

https://phabricator.endlessm.com/T22258