eos-ostree-remotes-config: Cleanup remotes.d config files
If there are config files in /etc/ostree/remotes.d or
/etc/flatpak/remotes.d that duplicate remotes in /ostree/repo/config,
then the repository can't be opened at all since ostree will error
immediately in that case. That makes handling them via flatpak or ostree
not an option, so the config handling is done here with python's
ConfigParser. In addition to setting the core.add-remotes-config-dir as
before, it now merges in any remotes.d config files and deletes them.

https://phabricator.endlessm.com/T22258