eos-update-flatpak-repos: Only operate on system installation
When calling the flatpak CLI, always specify --system to ensure that
flatpak doesn't manage remotes in the user's installation. This is
unlikely to make much difference in normal operation when the service is
running as root and /root/.local/share/flatpak is probably empty.
However, it's more robust to specify that the system installation is
operated on and allows running via sudo when HOME is set the calling
user's home directory and would pick up their ~/.local/share/flatpak
installation.

https://phabricator.endlessm.com/T22258