eos-config-journal: set correct permissions on /var/log/journal
This is as suggested by `man systemd-journal`:

> On systems where /var/log/journal/ does not exist yet but where
> persistent logging is desired (and the default journald.conf is used),
> it is sufficient to create the directory, and ensure it has the
> correct access modes and ownership:

    mkdir -p /var/log/journal
    systemd-tmpfiles --create --prefix /var/log/journal

Without this, /var/log/journal is owned by root:root, which prevents
administrators (who are members of 'adm' but not of 'root') from being
able to see privileged parts of the journal without `sudo`.

An alternative approach would be to order this job before
`systemd-tmpfiles-setup.service`, but there's no particular need to
block the boot process on creating this directory.

https://phabricator.endlessm.com/T22687