endless-repartition: stop creating swap partitions
Overall, we achieve better results with zram than with disk
swap partitions, so we are no longer going to create a swap
partition on first boot.

At the moment, if an existing user has already completed
repartitioning on an older OS version that created a swap partition,
we do not yet have a mechanism to remove it to reclaim the disk space,
and we still blindly disable all swap partitions when enabling zram.

https://phabricator.endlessm.com/T21758