eos-repartition-mbr: read exactly 8 bytes of backup GPT
Previously we would see the following warning on a successful run:

> eos-repartition-mbr: line 78: warning: command substitution: ignored null byte in input

This was because we read a full 512-byte sector at the start of where we
thought the GPT should be, and that sector would contain null bytes. But
we are only interested in 8 bytes which we expect to be the string "EFI
PART". We can avoid this warning by telling dd to read exactly 8 bytes.