eos-repartition-mbr: don't try to zero GPT headers
sfdisk now does this automatically, since
https://github.com/karelzak/util-linux/commit/5635d19:

> always (except --wipe=never) wipe old partition tables

with the nuance that it only wipes the "EFI PART" bytes of the table,
not the whole sector, which is fine. So:

* adjust the test accordingly
* drop our logic to try to do the same thing (which fails if sfdisk has
  this behaviour, because the EFI PART in the backup GPT is zeroed)
* add a versioned dependency on util-linux

I do not understand how https://phabricator.endlessm.com/T22665 worked
when I QAed it since we have a new-enough sfdisk in eos3.4. We should
backport this change for 3.4.2.