eos-update-flatpak-repos: rename/clean up ostree refs as flatpaks are migrated
Previous versions of this script left behind local ostree refs to the old
origin/branch of the apps it had migrated. This meant that once flatpak had
upgraded the app and pulled the new ref from the network, the old one was never
cleaned up, leaving behind the commit and disk space associated with the old
version of the app.

To both fix this problem, and prevent it happening again, we build a list of
matching ostree refs which should shadow the installed flatpaks we are migrating,
and copy the ref to the new name if the app branch or origin is changed during
the migration. The remaining refs in our list at the end of the process have
either just been migrated, or were migrated in the past, and are therefore
deleted.

https://phabricator.endlessm.com/T22763