eos-update-flatpak-repos: make kind and origin mandatory
When migrating flatpaks, make kind and the old origin mandatory to
ensure that we never accidentally match ostree refs which do not
correspond to flatpaks. As the migrations are performed in order
and the flatpak caches updated after each change, the correct
eos-sdk origin can be specified on the South-East Asia subpaths
migration, because the eos-runtimes origin change will be completed
beforehand.