eos-find-orphaned-runtimes: new helper script
This is a precursor to a comprehensive script to help fix users
who are stuck and cannot upgrade.

This script finds all the flatpak runtimes that are installed
in the system, and then iterates through all installed flatpak apps
to find all the runtimes (including extensions) that are needed
for these apps.  Any runtimes that are not found to be needed
are printed out.

https://phabricator.endlessm.com/T22791