eos-update-flatpak-repos: remove OSTree leaked refs for eos3
Due to a bug in eos-updater (T23385), upgrades from OS version 3.3 which
follow the checkpoint to move from the eos3 to the eos3a branch from
version 3.4 onwards do not remove the old eos:os/.../eos3 ref which
points to the last 3.3 release. This means that even after the 3.3 OS is
undeployed (after a second 3.4 update) the objects are kept around in
the OSTree repo - approx 3GB of wasted disk! Although we can fix the
updater bug for the next checkpoint, any system that has already
upgraded, and anyone else who uses the (still buggy) 3.3 updater to
follow the checkpoint, will have the leaked eos3 ref.

https://phabricator.endlessm.com/T23410