udev/dracut: enable BFQ by default on disk devices
Set the scheduler to BFQ on block devices with a device type of disk, and add
the bfq module to the initrd so this udev rule works from early boot. This will
only impact SCSI devices on Endless systems, where we have switched the default
of the SCSI disk devices to the new MQ scheduler framework, which is a
pre-requisite of BFQ. Any block devices which don't currently use the MQ
framework can't access BFQ, such as MMC under older kernels, in which case this
action is a no-op and the current in-kernel default of CFQ remains.

https://phabricator.endlessm.com/T23725
(cherry picked from commit a290c97d97befd8d15e586dd09b7e32383ba48b4)
Signed-off-by: Robert McQueen <rob@endlessm.com>