eos-update-flatpak-repos: check for and handle already-migrated apps
Changes of origin don't change the deployed path on disk, and our
origin-changing code is idempotent, so we only need to check if something is
installed in the case the branch is being changed. If the ref we are migrating
to already exists, uninstall the old one.

https://phabricator.endlessm.com/T21141
(cherry picked from commit 611271d75b5ee1b0e5568f4d7a2f8093960e25e1)
Signed-off-by: Robert McQueen <rob@endlessm.com>