eos-update-flatpak-repos: drop support for migrations that do not change the ref
Previously we had some in-place transformations of the deploy such as modifying
the subpaths, but these are no longer present. We are keeping the back catalog
of refs that do change the branch and origin so that we can clean up the leaked
ostree refs from the 3.3 version. The subpath modifications didn't change any
refs so didn't cause a leak, so they could be deleted after the eos3.4 (eos3a)
branch checkpoint.

(cherry picked from commit 0a8ca3614a60e0c591844459a14e954ca491de3a)
Signed-off-by: Robert McQueen <rob@endlessm.com>