eos-update-flatpak-repos: copy ostree ref as first operation
Move the copy of the ref from the old name to the new name to the start. This
will allow us to later use flatpak to operate on the local ref as if it had
been pulled from the real remote, but without requiring any network.

(cherry picked from commit fc6cd2e5f1d1ebd5892bff7ab43744202ba025d7)
Signed-off-by: Robert McQueen <rob@endlessm.com>