eos-update-flatpak-repos: use Flatpak API to change branch names
Now that we copy the OSTree refs ahead of any Flatpak migrations, we can use a
NO_PULL deploy to install the Flatpak with the new branch name, as it will be
created in a new folder. Then we can unconditionally (whether or not the new
one was already installed or not) remove the old one. At a slight performance
cost (as the app will now be redeployed), the hair-raising update branch
function which moved folders around and monkey patched the exported files
behind Flatpak's back can be removed.

We now only directly access the deploy file in the case we are modifying the
origin, but this code path is avoided if we've just done redeploy from the new
ref, because it will already have the new origin. As this is the last and only
action performed outside of the Flatpak API, we don't need to drop the cache or
refresh the ref.

https://phabricator.endlessm.com/T22902
(cherry picked from commit 565d396f11d8ad9d3a3b2eb1d8b406bdcfa4647c)
Signed-off-by: Robert McQueen <rob@endlessm.com>