eos-update-flatpak-repos: support app renaming
Add a 'new-name' key (which can only be used in combination with 'name', not
'prefix') to migrate an app from one ID to another. Relying on the refactoring
to implement branch changes by copying the ref, and asking Flatpak to do a
re-deploy, this works in broadly the same manner.

We create a commit to change the ref to the new name, but introduce a new
function to modify the OSTree MutableTree before the new tree is committed. We
modify the Flatpak metadata to reflect the new ID, and add the old app ID as
a permitted D-Bus name, and .service files are not renamed so that any app
interfaces are not changed. Desktop files are renamed, references to other
exported assets such as icons and splash screens are updated, and if enabled,
D-Bus activation is disabled because there is no longer a match between the
desktop file name and any D-Bus name implemented in the app. Other exported
files are renamed to match the new ID so that they are correctly exported by
Flatpak.

At the moment, Extra Data flatpaks are excluded, because it is not possible to
re-deploy them without connectivity, and any migration that matches more than
one ref (for example, an app with extensions) is also prohibited as we do not
recursively re-map extension names in the metadata. I do not believe our third
party app bundling system had the capacity to produce such apps.

https://phabricator.endlessm.com/T23819
(cherry picked from commit 471f5411fe2757a291a7ba5bf70f1adba42836fe)
Signed-off-by: Robert McQueen <rob@endlessm.com>