eos-update-flatpak-repos: build new commit with correct metadata
Flatpak emits a warning (or, if collection IDs are enabled, a fatal error)
deploying an app or runtime if the branch on xa.ref binding attribute does not
match the app which is being deployed. Import the copy_commit function from
ostree-upload-server, which is a Python equivalent to `flatpak
build-commit-from`. It takes the existing directory tree and copies most commit
metadata, but sets the ostree and Flatpak ref bindings correctly. It can also
set the collection ID binding if set, but we do not currently pass one, so this
is removed, because we do not have the right key to sign the imported ref, so
we don't want to export it either.

(cherry picked from commit a7fb401a2380842bbcd6056832ee9cf23b2a0d25)
Signed-off-by: Robert McQueen <rob@endlessm.com>