eos-update-flatpak-repos: update X-Flatpak-RenamedFrom
Our Flatpak is patched to propagate a previous-ids property from old
to new deploy file when deploying an app update, and to add all previous
IDs to the X-Flatpak-RenamedFrom key in the process.

However, in this script we only set previous-ids after deploying the
patched version of the app (under its new name). In any case, as far as
Flatpak is concerned, this is a wholly new app being installed, so there
would be no previous deployment to carry previous-ids forward from.

Since we are patching the .desktop files anyway while doing the
migration, we can just implement the same logic to update
X-Flatpak-RenamedFrom here, and add a test for this and other .desktop
file mangling.

https://phabricator.endlessm.com/T23845
(cherry picked from commit 610485aa86576b9b3923d56b027f20d706c74fee)
Signed-off-by: Robert McQueen <rob@endlessm.com>