Don't use bare 'except:'
The difference between 'except:' and 'except Exception:' is that
SystemExit (raised by exit(n)) and KeyboardInterrupt (raised by ^C) are
not instances of Exception, so are caught by the former but not the
latter. Some versions of flake8 warn about using bare 'except:' since
if misused it can lead to programs that can't be interrupted, and which
don't terminate when they call exit().

In all cases here, we're re-raising the exception, so they were actually
all safe. In the tests, we don't need to dump debugging information when
the tests are interrupted, so we use 'except Exception:'. In
eos-update-flatpak-repos, we do want to cleanly abort any open
transaction on *any* exception, including SystemExit or
KeyboardInterrupt. ostree_repo_abort_transaction() is a no-op when no
transaction is open, so we can move it to a 'finally:' block to give the
behaviour we want.

(cherry picked from commit 0042f1238831e84ec208a7400e108eafd357e77c)
Signed-off-by: Robert McQueen <rob@endlessm.com>