eos-update-flatpak-repos: stuff vendor prefix into previous-ids
In an earlier commit, we moved any .desktop files in a vendor/
subdirectory up to the toplevel applications/ directory, adding
$vendor-$old_basename to X-Flatpak-RenamedFrom in the process. However,
we need this to be preserved across app updates. The upstream version
will not have this filename in this key; and our downstream Flatpak
logic to preserve X-Flatpak-RenamedFrom across app updates is based on
the previous-ids key in the deploy metadata. This only includes the old
Flatpak app ID, not the old names of .desktop files.

We can trick it into doing what we want by pretending that (for example)
org.kde.klickety has 2 previous-ids, namely org.kde.Ksame (the real old
app ID) and kde4-org.kde.Ksame. The latter will cause
"kde4-org.kde.Ksame.desktop" to be added to X-Flatpak-RenamedFrom, which
is what we want. The chances are effectively zero that this will clash
with a real app ID, and there will be no existing data in
.var/app/kde4-org.kde.Ksame so this should have no ill effects.

Not pretty, but works.

https://phabricator.endlessm.com/T23856
(cherry picked from commit 399aeac842af6942d45a92ca7f3dce9145f763c9)
Signed-off-by: Robert McQueen <rob@endlessm.com>