eos-update-flatpak-repos: update xa.metadata on commit
The more recent apps such as org.processing.App built with flatpak-builder have
a copy of the Flatpak metadata file in the OSTree commit metadata, under the
xa.metadata key. If set, Flatpak ensures that it matches the contents of the
file, so we reset it with the metadata we've added to our new commit, otherwise
these apps cannot be deployed.