eos-update-flatpak-repos: disable OSTree fsyncs
Explained more verbosely with a comment in the code, we only produce commits
and refs before installing them with Flatpak, an operation which concludes with
a syncfs() call, so we can avoid the syncs built into OSTree. Deleting a ref
(which we do at the end) is an atomic unlink() so is unaffected by this.

One other change is made for safety: in case a commit has been made previously
but has not been installed, we migrate the OSTree commit from the old ref to
the new ref, overwriting any previously-attempted migration for this app.