eos-update-flatpak-repos: handle case where app name does not change
Fix a regression introduced in 399aeac8 which assumes that any app having its
commit copied is also having its name changed. In the case that an app is
migrated from eos-apps to flathub without a name change (such as Stellarium) we
are only changing the branch name and origin, so the rewriting code path does
not run and vendor_prefixes is not set.