eos-update-flatpak-repos: handle multiple renames to the same name
The code introduced in 471f5411 didn't correctly handle the case where we
rename two separate applications (in the same run) to the same name. The
expected behaviour is that when renaming the 2nd app, we detect that the target
is already installed, so we uninstall it and add it's name to the previous IDs
of the already renamed app. The detection was not working due to the first
rename not updating the cached list of flatpak refs which is used to find
whether an app is installed.

(cherry picked from commit e33dfcdbfe8cf5a3ce261ea013bc9e64ff260ad8)