eos-update-flatpak-repos: do not migrate Freecell and Solitaire for now
These apps were represented previously as two separate apps, although they are
actually just copies of the same Aisleriot app. Migrating/renaming both of them
has the result that the .desktop file doesn't contain one or other previous ID,
meaning that shell icon grids will not be updated correctly. Accepting that for
a while, we might see Freecell, Solitaire and Aisleriot together in the app
center, we're going to ship Aisleriot moving forwards but delay the rename
until we've come up with a way to handle the previous IDs in the desktop file
better.