eos-update-flatpak-repos: don't log twice to journal
logging.basicConfig(), if called without any handlers (and with no
handlers previously configured), adds a logging.StreamHandler for stderr.
After calling it, we previously unconditionally configured a
JournalHandler to send log output to the systemd journal.

When run interactively, stderr isn't connected to the journal; but when
run for real after an upgrade, it is. This meant we logged everything to
the journal twice, making it rather hard to read. As it happens, the
prefix on each line was the same length, which helped a bit:

Oct 08 11:22:10 endless eos-update-flatpak-repos[332]: INFO:root:Found eos-apps:app/org.stellarium.Stellarium/x86_64/eos3 to migrate to flathub:app/org.stellarium.Stellarium/x86_64/stable
Oct 08 11:22:10 endless /usr/sbin/eos-update-flatpak-repos[332]: Found eos-apps:app/org.stellarium.Stellarium/x86_64/eos3 to migrate to flathub:app/org.stellarium.Stellarium/x86_64/stable

But it's easy enough to avoid this, by only logging to stderr if it is a
tty. This does the wrong thing if someone explicitly redirects stderr to
a file, but in practice no-one will do this.

https://phabricator.endlessm.com/T24289