eos-reclaim-swap: avoid cylinder alignment for DOS partition tables
For DOS partition tables, the msdos_partition_align() function called
in the disk.maximizePartition() codepath was modifying the partition
start, aligning it based on BIOS cylinder calculations. Drop that
constraint so that the sector will be unchanged.

This was causing the new partition to be written with a different start
sector (144585 instead of 131072), but here we are only trying to modify
the end. Add a sanity check to prevent that from happening again for
any reason.

https://phabricator.endlessm.com/T24454