Add eos-factory-reset scripts and systemd units
This adds the following components

- A script to trigger the factory reset, which basically creates a flag
file with the factory reset stages to execute.
- This adds a unit file to be run under basic.target, before multi-user.target
starts, so that's the reason it's WantedBy the basic.target, which is the
latest step before multi-user.target where we have most of the system booted up
(filesystem and essential system services), this approach is very similar to
system-update or eos-firstboot. It is enabled by default but it only runs when
the flag file exists.
- A series of helper scripts that perform the actual reset.