factory-reset: Add wrapper to schedule each stage
Add the eos-factory-reset program to schedule each stage of the factory
reset process to the next boot. Each stage can be enabled independently,
and adding a new stage is simply a matter of its name to the STAGES
list, in addition to shipping the program and unit file named
"eos-factory-reset-<stage_name>.service".

eos-factory-reset does not need to be run as root, as policykit will
take care of asking for authentication if needed.

If enabling one of the requested stages fails, it disables all the
stages that were requested to be enabled on that invocation.  Note that
this is not exactly transactional, as if the user tries to enable stages
A then B, and enabling A fails, both stages will be disabled
unconditionally. So if B was enabled before the program invocation it
will end-up disabled in this example. That should not be a problem for
the use-cases we are designing this for, though.

There is also the command-line option '--disable' to disable all
factory-reset stages know to this tool, completely un-scheduling a
factory reset.

In addition to printing status on the terminal, logs are also sent to
the journal In the future if we add clearing logs as part of the factory
reset process, we will need to save a copy of the pre-reset journal
(probably in a location that will be cleared automatically later) if we
want to be able to diagnose problems after the fact.

https://phabricator.endlessm.com/T24613