fallback-fb-setup: Don't clear the fb if the VT is in text mode
The virtual terminal may be put in text mode during boot if the previous
boot failed (so grub increases the kernel log level for the subsequent
boot), or if the user pressed ESC during the boot process. If the VT is
in text mode, when the boot process is finished but GDM has not started
yet, clearing the fallback framebuffer will leave the user for several
seconds looking at either some garbled output or a black screen, none of
which look great. Also, it does not help having a clean shutdown in this
case, as the messages in the console will still be displayed when the
user's session is terminated. Finally, it may make it more difficult to
gather debug information from the user in case errors are printed to the
console as the period during boot that the errors will be available will
be much shorter.

This commits makes fallback-fb-setup only clear the fallback framebuffer
if the active VT is in graphics mode.

https://phabricator.endlessm.com/T24682