eos-preset: disable speech-dispatcherd.service
It fails to start due to the lack of a system-wide PulseAudio session:

  pulse.c: pa_simple_new() failed: Connection refused

According to README.Debian, running a system-wide instance of this
daemon is not a recommended configuration, and debian/rules uses
"dh_systemd_enable --no-enable" to disable it by default on Debian
systems. Disable it here, too. (Previously, the package did not include
a .service file, and the init script checked an off-by-default flag in
/etc/default/speech-dispatcher.)

https://phabricator.endlessm.com/T25605